updated model for frontend

+++ controllers/item.js
@@ -1,17 +1,18 @@
 var itemModel = require('../models/Item');
+var raffleModel = require('../models/raffle');
 var ObjectId = require("mongodb").ObjectID;
 var fs = require('fs');
 
 module.exports = {
     create: function (request, response, next) {
+
         if (request.body.name &&
             request.body.description &&
             request.body.createdBy &&
-            request.file
+            request.body.img
         ) {
             itemModel.create({
-                'img.data': fs.readFileSync(request.file.path),
-                'img.contentType': 'image/png',
+                img: request.body.img,
                 createdBy: ObjectId(request.body.createdBy),
                 created: Date.now(),
                 name: request.body.name,
@@ -39,9 +40,9 @@ module.exports = {
                 }
             });
     },
-    allItem: function (request, response, next) {
-        itemModel.find({}, {
-                img: 0
+    findByCreatedBy: function (request, response, next) {
+        itemModel.find({
+                createdBy: new ObjectId(request.params.id)
             })
             .exec(function (err, item) {
                 if (err) {
@@ -57,13 +58,15 @@ module.exports = {
     updateById: function (request, response) {
         if (
             request.body.name &&
-            request.body.description
+            request.body.description &&
+            request.body.img
         ) {
             itemModel.updateOne({
                 _id: new ObjectId(request.params.id)
             }, {
                 name: request.body.name,
-                description: request.body.description
+                description: request.body.description,
+                img: request.body.img
             }).exec(function (err, item) {
                 if (err) {
                     response.send(err);
@@ -77,14 +80,26 @@ module.exports = {
         }
     },
     deleteById: function (request, response) {
-        itemModel.deleteOne({
-            _id: new ObjectId(request.params.id)
-        }).exec(function (err) {
-            if (err) {
-                response.send(err);
-            } else {
-                response.send("Successful.");
-            }
-        });
+        raffleModel.find({
+                item: new ObjectId(request.params.id)
+            })
+            .exec(function (err, raffle) {
+                if (err) {
+                    response.send(err);
+                } else if (raffle.length <= 0) {
+                    itemModel.deleteOne({
+                        _id: new ObjectId(request.params.id)
+                    }).exec(function (err) {
+                        if (err) {
+                            response.send(err);
+                        } else {
+                            response.send("Successful.");
+                        }
+                    });
+                } else {
+                    response.send("Delete Raffle first then Item.");
+                }
+            });
+
     }
 };
